categories_app/api/serializers.py
from rest_framework import serializers
from categories_app.models import Category


class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ["id", "name", "description", "parent"]
