requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
mysqlclient==2.2.0
pillow==11.3.0
djangorestframework>=3.16
